test(core): split runtime auto polyfill tests
